requirements.txt
SQLAlchemy
mysql-python
python-odesk
